mdss: ppp: Reset parameters passed into mdp3_start_ppp
mdp3_start_ppp is manipulating some of the parameters passed in.
Within a sw workaround to a hw issue, a blit request can be broken
down and create a need for multiple calls to mdp3_start_ppp. When
this happens, the parameters passed in will be modified and a
snowball effect takes place. To fix this issue we simply reset
parameters before next call to mdp3_start_ppp.

Change-Id: I29888b5f2f07391d3502ae1478068d7bf12805c3
Signed-off-by: Terence Hampson <thampson@codeaurora.org>