msm: mdss: allow asynchronous acquire fence waits
Add optional flag to allow MDP driver to asynchronously wait for fences
instead of doing it at beginning of every commit. This allows the wait
to happen much later in the programming sequence so that, we can use
that time to do everything that is needed and have a smaller window
between the wait for fences and the next frame update.

Change-Id: Ib4f7601ed22670ba235ba86fe3389a5d54d685a9
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>